% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_env.R
\name{import_env}
\alias{import_env}
\title{Import environmental base data}
\usage{
import_env(env_dir = NULL, sites = NULL, save = FALSE, save_dwnld = FALSE, save_dir = getwd())
}
\arguments{
\item{env_dir}{Path to local file containing environmental data;
NULL if downloading directly from EDE; Default = NULL.}

\item{sites}{Vector of site ids to filter on.}

\item{save}{Specifies if imported environmental data should be saved
as rds file (for future use); Default = FALSE.}

\item{save_dwnld}{Specifies if downloaded biology data should be saved.}

\item{save_dir}{Path to folder where imported environmental data is to be saved;
Default = Current working directory.}
}
\value{
Tibble containing environmental data
}
\description{
The \code{import_env} function imports macroinvertebrate sampling data from the Environment Agency's Ecology and Fish Data Explorer. The data can either be downloaded from https://environment.data.gov.uk/ecology-fish/downloads/INV_OPEN_DATA.zip or read in from a local csv or rds file. The data can be optionally filtered by site ID.
}
\details{
If saving a copy of the downloaded data, the name of the rds file is hard-wired to: INV_OPEN_DATA_SITES_ALL.rds. If saving after filtering on site, the name of the rds file is hard-wired to: INV_OPEN_DATA_SITE_F.rds.

Downloaded raw data files (in .csv and .zip format) will be automatically removed from the working directory following completed execution of the function.

The function will modify the output from EDE, renaming \code{SITE_ID} to \code{biol_site_id} (standardised column header for biology sites).
}
\examples{
# Download data for all sites and save as .rds file for future use:
  # import_env(save_dwnld = TRUE, save_dir = "mydata")

# Read in local .rds file and filter on selected sites:
  # import_env(env_dir = "mydata/INV_OPEN_DATA_SITES_ALL.rds",
  #                sites = c("34310", "34343"))

# Read in local .csv file and filter on selected sites:
  # import_env(env_dir = "mydata/INV_OPEN_DATA_SITES_ALL.csv",
  #               sites = c("34310", "34343"))
}
