% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rhs.R
\name{import_rhs}
\alias{import_rhs}
\title{Import RHS data}
\usage{
import_rhs(rhs_dir = NULL, surveys = NULL, save = FALSE, save_dwnld = FALSE, save_dir = getwd())
}
\arguments{
\item{rhs_dir}{Path to local file containing RHS data.}

\item{surveys}{Vector of survey ids to filter on.}

\item{save}{Specifies if imported RHS data should be saved to working directory
as rds file (for future use); Default = FALSE.}

\item{save_dwnld}{Specifies if downloaded RHS data should be saved.}

\item{save_dir}{Path to folder where RHS data is to be saved;
Default = Current working directory.}
}
\value{
Tibble containing RHS data.
}
\description{
The \code{import_rhs} function imports River Habitat Survey (RHS) data. The data can either be downloaded from from data.gov.uk (\url{https://environment.data.gov.uk/portalstg/sharing/rest/content/items/b82d3ef3750d49f6917fff02b9341d68/data}) or read in from a local xlsx or rds file. Data can be optionally filtered by survey ID.
}
\details{
If saving a copy of the downloaded data, the name of the rds file is hard-wired to: RHS_survey_summary_ALL.rds. If saving after filtering on site, the name of the rds file is hard-wired to: RHS_survey_summary_F.rds.

Downloaded raw data files (in xlsx and zip format) are automatically removed from the working directory following completed execution of the function.
}
\examples{
# Download data for all surveys and save as .rds file for future use:
# import_rhs(save_dwnld = TRUE, save_dir = "mydata")

# Read in local .rds file and filter on selected dates:
# import_env(rhs_dir = "mydata/RHS_survey_summary_ALL.rds",
#                  surveys = c("34310", "34343"))

# Read in local .xlsx file and filter on selected sites:
# import_rhs(rhs_dir = "mydata/RHS_ALL.xlsx",
#                  surveys = c("34310", "34343"))
}
