% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_taxa.R
\name{import_taxa}
\alias{import_taxa}
\title{Importing macroinvertebrate abundance data from the EA Ecology and Fish Data Explorer}
\usage{
import_taxa(sites = NULL,
            samples = NULL,
            save = FALSE)
}
\arguments{
\item{sites}{Vector of site ids to filter by. Default = NULL}

\item{samples}{Vector of sample ids to filter by. Default = NULL}

\item{save}{Specifies whether (\code{TRUE}) or not (\code{FALSE}) the filtered data should be saved as an rds file (for future use, or audit trail). Default = \code{FALSE}.}
}
\value{
A tibble containing the imported macroinvertebrate taxonomic abundance data.
}
\description{
The import_taxa function imports macroinvertebrate taxon abundance information from the Environment Agency's Ecology and Fish Data Explorer (EDE). The function downloads and merges two files: one listing taxon abundances by sample, and one with details of each taxon (e.g. full taxon name, taxonomic rank). The data can be optionally filtered by site and/or sample ID, and the filtered data saved as a .rds file.
}
\details{
If saving a copy of the filtered taxa data, the name of the rds file is hard-wired to \code{INV_OPEN_DATA_TAXA_F.RDS}.

Downloaded raw data files will be automatically removed from the working directory following completed execution of the function.

The function automatically modifies the output from EDE, renaming "SITE_ID" to "biol_site_id" (\code{hetoolkit}'s standardised column header for biology site ids).

Please note that that approaches to sample processing, such as abundance recording, have changed over time. For details, please see page 5 of \href{https://environment.data.gov.uk/api/file/download?fileDataSetId=66c79b9a-c855-4b65-a894-0668bbbd1843&fileName=Freshwater_river_macroinvertebrate_surveys_(Biosys).pdf}{this pdf document}
}
\examples{

# Import data from EDE for specific sites
taxa_data <- import_taxa(sites = c("34310", "34343"))

# Import data and match on other macroinvertebrate data
biol_data <- hetoolkit::import_inv(sites = c("34310", "34343"))
sample_IDs <- unique(biol_data$SAMPLE_ID)

taxa_data <- import_taxa(samples = sample_IDs)

full_data <- taxa_data \%>\%
  dplyr::left_join(biol_data)

}
