% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hev.R
\name{plot_hev}
\alias{plot_hev}
\title{Producing time series plots of biology and flow data}
\usage{
plot_hev(data, date_col, flow_stat, biol_metric, multiplot = TRUE, save = FALSE, save_dir = getwd(), clr_by = NULL)
}
\arguments{
\item{data}{Data to plot.}

\item{date_col}{Name of column containing time variable for x-axis.}

\item{flow_stat}{Names of up to two flow statistics of interest.}

\item{biol_metric}{Names of up to four biology metrics of interest.}

\item{multiplot}{If TRUE, a multiplot of up to four biology metrics is produced;
if FALSE, a single plot is produced for each biology metric. Default = TRUE.}

\item{save}{Specifies if outputs should also be saved as png files. Default = FALSE.}

\item{save_dir}{Path to folder where png files are to be saved;
Default = Current working directory.}

\item{clr_by}{Selected variable by which to colour biol_metric points. Default = NULL.}
}
\value{
Depending on the multiplot argument, the function returns either a single ggplot
object, or a list of ggplot objects.
}
\description{
Intended for plotting data for one site at a time; data may need to be subset or filtered
by user prior to plotting.

The function generates a ggplot object, that is optionally saved as a .png file.
If multiplot = TRUE, the png file is named "Multi_Plot.png". If multiplot = FALSE, a separate png file is produced for each biology metric, named "biolmetric_Plot.png".
}
\examples{
## Generate hev plots for four biology metrics at one site of interest
# plot_hev(data = subset(all_data, biol_site_id = "100582"),
#            date_col = "SAMPLE_DATE",
#            flow_stat = c("Q95z", "Q10z"),
#            biol_metric = c("LIFE_F_OE", "WHPT_ASPT_OE", "WHPT_NTAXA_OE", "PSI_OE"),
#            multiplot = TRUE)
}
