% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_hev.R
\name{shiny_hev}
\alias{shiny_hev}
\title{Produces an interactive time series plot of biology and flow data for one site of interest.}
\usage{
shiny_hev(data, sites_col, date_col, flow_stat, biol_metric, multiplot = TRUE)
}
\arguments{
\item{data}{Data frame, containing the data to plot.}

\item{sites_col}{Name of variable defining site ID.}

\item{date_col}{Name of column containing time variable.}

\item{flow_stat}{Flow statistics of interest; two should be chosen.}

\item{biol_metric}{Biology metrics of interest; up to four can be chosen.}

\item{multiplot}{If TRUE: a multiplot of up to four biology metrics is displayed;
If FALSE: a single  plot is displayed.  Default = TRUE.}
}
\value{
Launches a shiny app containing an interactive time series plot of biology and flow data.
}
\description{
The shiny_hev function launches a shiny app of:
\itemize{
\item If multiplot = TRUE, a multiplot plotly plot with:
(i)	a drop-down list to select the site of interest.
\item If multiplot = FALSE, a plotly plot with:
(i)	a drop-down list to select the site of interest, and
(ii)	a drop-down list to select which biol_metric to plot.
}
}
\examples{
## Multi-panel plot

# shiny_hev(data = all_data,
#            sites_col = "biol_site_id",
#            date_col = "Year",
#            flow_stat = c("Q10z", "Q95z"),
#            biol_metric = c("WHPT_ASPT_OE", "LIFE_F_OE"),
#            multiplot = TRUE)
}
