% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_inv.R
\name{import_inv}
\alias{import_inv}
\title{Import macroinvertebrate data in Ecology Data Explorer (EDE) format}
\usage{
import_inv(source = "parquet", sites = NULL, start_date = NULL, end_date = NULL, save = FALSE, save_dwnld = FALSE, save_dir = getwd()biol_dir = NULL)
}
\arguments{
\item{source}{File format for the download from EDE OR path to local .csv or .rds file containing biology data. Default = "parquet".}

\item{sites}{Vector of site ids to filter by.}

\item{start_date}{Start date for data extraction (YYYY-MM-DD format). Default = NULL.}

\item{end_date}{End date for data extraction (YYYY-MM-DD format). Default = NULL.}

\item{save}{Specifies if filtered biology data should be saved as rds file (for future use). Default = FALSE.}

\item{save_dwnld}{Specifies whether or not downloaded biology data should be saved. Default = FALSE.}

\item{save_dir}{Path to folder where biology data is to be saved. Default = Current working directory.}

\item{biol_dir}{Deprecated. Path to local .csv or .rds file containing biology data; or NULL to download directly from EDE. Default = NULL.}
}
\value{
Tibble containing biology data
}
\description{
The \code{import_inv} function imports macroinvertebrate sampling data from the Environment Agency's Ecology and Fish Data Explorer. The data can either be downloaded from \url{https://environment.data.gov.uk/ecology-fish/downloads/INV_OPEN_DATA.zip} or read in from a local csv or rds file. The data can be optionally filtered by site ID and sample date.
}
\details{
If saving a copy of the downloaded data, the name of the rds file is hard-wired to: INV_OPEN_DATA_METRICS_ALL.RDS. If saving after filtering on site or date, the name of the rds file is hard-wired to: INV_OPEN_DATA_METRICS_F.RDS.

Downloaded raw data files (in .csv and .parquet format) will be automatically removed from the working directory following completed execution of the function.

The function will modify the output from EDE, renaming "SITE_ID" as "biol_site_id" (standardised column header for biology sites).
}
\examples{
# Download data for all sites and save as .rds file for future use:
# import_inv(save_dwnld = TRUE, save_dir = "mydata")

# Read in local .rds file and filter on selected sites and dates:
# import_inv(biol_dir = "mydata/INV_OPEN_DATA_METRICS_ALL.rds",
#                  sites = c("34310", "34343"),
#                  start_date = "1995-01-01",
#                  end_date = Sys.Date())

# Read in local .csv file, filter on selected sites, and save results as rds file:
# import_inv(biol_dir = "mydata/INV_OPEN_DATA_METRICS.csv",
#                  sites = c("34310", "34343"),
#                  save = TRUE)
}
