% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_flow.R
\name{impute_flow}
\alias{impute_flow}
\title{Impute missing flow data.}
\usage{
impute_flow(data, site_col = " flow_site_id", date_col = "date", flow_col = "flow", method = "linear", donor = NULL)
}
\arguments{
\item{data}{A tibble or data frame containing the data to be imputed.}

\item{site_col}{Name of column in data containing unique flow site id. Default = "flow_site_id". Site ids are coerced to a character vector.}

\item{date_col}{Name of column in data containing date of flow record. Default = "date".  Dates must be in “yyyy-mm-dd” format.}

\item{flow_col}{Name of column in data containing the measured  flow data. Default = "flow".}

\item{method}{Imputation method: "linear" (default), "exponential" or "equipercentile".}

\item{donor}{– A tibble or data frame with at least two columns: the first a list of flow sites requiring imputation, and the second a list of paired donor sites. Subsequent columns are ignored. Default = NULL. Only used when method = "equipercentile".}
}
\value{
A tibble containing the imputed flow data. The data are arranged in long format, with the following columns:
\itemize{
\item flow_site_id (unique flow site id)
\item date (of flow record)
\item imputed (flag  indicating whether each flow value is original (0) or imputed (1))
\item donor_site (id of donor site used for imputation, if "equipercentile" method used)
\item donor_flow (measured flow at donor site on that date, if "equipercentile" method used)
\item any other columns in the input dataset are automatically pulled through and joined to the output data table (e.g. the ‘input’ and ‘quality’ columns from the import_flow() function).
}
}
\description{
This function infills missing records in daily flow time series for one or more sites (gauging stations) using either interpolation or an equipercentile  method. Imputation of missing flow data can improve the later estimation of flow statistics using the calc_flowstats() function and aid the visualisation of hydro-ecological relationships using the plot_hev() function.
}
\details{
This function is intended for imputing gauged daily flow data only; it cannot be used to impute sub-daily data, and is not designed for data on coarser time steps (e.g. 10-daily or monthly).

The function offers three imputation methods: linear interpolation, exponential interpolation and an equipercentile algorithm.
}
\examples{
## impute flow statistics using 'linear' method
impute_flow(data_impute,
            site_col = "flow_site_id",
            date_col = "date",
            flow_col = "flow",
            method = "linear")

## impute flow statistics using 'exponential' method
impute_flow(data_impute,
            site_col = "flow_site_id",
            date_col = "date",
            flow_col = "flow",
            method = "exponential")

impute flow statistics using 'equipercentile' method, without specifying the donor station to be used
impute_flow(data_equipercentile,
            site_col = "flow_site_id",
            date_col = "date",
            flow_col = "flow",
            method = "equipercentile")

impute flow statistics using 'equipercentile' method, without specifying the donor station to be used
impute_flow(data_equipercentile,
            site_col = "flow_site_id",
            date_col = "date",
            flow_col = "flow",
            method = "equipercentile",
            donor = donor_data)

}
