% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rngflows.R
\name{plot_rngflows}
\alias{plot_rngflows}
\title{Visualise the range of flow conditions experienced historically}
\usage{
plot_rngflows (data, flow_stats, biol_metric, wrap_by = NULL, plotly = FALSE, label = NULL)
}
\arguments{
\item{data}{Name of data frame or tibble containing the data to be analysed, such as that produced by the join_he function.}

\item{flow_stats}{Vector of two flow statistics to be plotted.}

\item{biol_metric}{Name of column containing biology data; NAs are interpreted as indicating the absence of a sample.}

\item{wrap_by}{Grouping variable to facet the plot (e.g. site). Default = NULL.}

\item{label}{Optional variable (e.g. date) to label points in plotly plot. Default = NULL.}

\item{plotly}{Logical value specifying whether or not to render the plot as an interactive plotly plot. Default = FALSE.}
}
\value{
A single ggplot2 or plotly object.
}
\description{
plot_rngflows generates a scatterplot for two flow variables (such as those produced by calc_flowstats) and overlays two convex hulls: one showing the full range of flow conditions experienced historically, and a second convex hull showing the range of flow conditions with associated biology records. This allows the user to assess to coverage of the biology data with respect to the historical range of flow conditions.
}
\examples{
## Example 1: Produce a single plotly plot, combining data for all sites
# plot_rngflows(data = all_data,
#               flow_stats = c("Q95z", "Q10z"),
#               biol_metric = "LIFE_F_OE",
#               wrap_by = NULL,
#               label = "Year")

## Example 2: Produce a faceted ggplot, showing data separately for each site
# plot_rngflows(data = all_data,
#               flow_stats = c("Q95z", "Q10z"),
#               biol_metric = "LIFE_F_OE",
#               wrap_by = "biol_site_id",
#               label = "Year")
}
